div. fixes
